renamed admin_login to login and the initial login to login1
