+++ login.php
@@ -1,101 +1,140 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: Hollyphat
- * Date: 10/17/2019
- * Time: 9:12 AM
- */
-
-session_start();
-if(isset($_GET['act']) && ($_GET['act'] == "logout")){
-    unset($_SESSION['hng-admin']);
-    header("location:login.php");
+<?php require 'classControllers/init.php';
+$interns = new Admins;
+if (isset($_POST['login'])) {
+  $email = $_POST["email"];
+  $password = $_POST["password"];
+  $login = $interns->adminLogin($email, $password);
+  $errors = '<p style="margin: 5px; padding: 5px 10px 5px 10px; background: #F7CFCF; color: #6A0E0D; width: 80% !important; text-align: center;">Wrong Credentials</p>';
 }
 
-require 'classControllers/init.php';
-$admin = new AdminClass();
 
 
-if(isset($_POST['login'])){
-    $email = $database->escape_string($_POST["email"]);
-    $password = $database->escape_string($_POST["password"]);
-
-    if($admin->login("$email","$password")){
-        $admin_info = $admin->getAdminDetails($email);
-        $_SESSION['hng-admin'] = $admin_info; //Set session for logged user
-        header("location:test_home.php"); //redirect to test home
-    }else{
-        $_SESSION['err_msg'] = "<div class='alert alert-danger'>Invalid login details!</div>"; //Set error message
-    }
-
-
-}
 ?>
 <!DOCTYPE html>
-<html>
+<html lang="en">
+
 <head>
-    <title>HNG i7</title>
-    <!-- Latest compiled and minified CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css" integrity="sha384-HSMxcRTRxnN+Bdg0JdbxYKrThecOKuH5zCYotlSAcp1+c8xmyTe9GYg1l9a69psu" crossorigin="anonymous">
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Admin Login</title>
+  <link rel="shortcut icon" href="https://res.cloudinary.com/dekillerj/image/upload/v1570648980/brand-logo.png" />
+  <link href="https://fonts.googleapis.com/css?family=Open+Sans&display=swap" rel="stylesheet" />
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp" crossorigin="anonymous" />
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.11.2/css/all.css" />
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+  <link rel="icon" type="img/png" href="images/hng-favicon.png">
+  <link rel="stylesheet" type="text/css" href="css/header-footer.css">
+  <link rel="stylesheet" href="css/join-intern.css">
+  <script type="text/JavaScript" src="https://ajax.googleapis.com/ajax/libs/jquery/1.10.1/jquery.min.js"></script>
+  <script>
+    function signup() {
 
-    <!-- Optional theme -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap-theme.min.css" integrity="sha384-6pzBo3FDv/PJ8r2KRkGHifhEocL+1X2rVCTTkUfGk7/0pbek5mMa1upzvWbrUbOZ" crossorigin="anonymous">
+      var email = $('#email').val();
+      var password = $('#password').val();
 
-    <!-- Latest compiled and minified JavaScript -->
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js" integrity="sha384-aJ21OjlMXNL5UyIl/XNwTMqvzeRMZH2w8c5cRVpzpU8Y5bApTppSuUkhZXN0VxHd" crossorigin="anonymous"></script>
+      $.post($login.
+          '?login=yes", {
+          email: email,
+          password: password
+        },
+        function(data) {
+          $('#result').html(data);
+        });
+    }
+  </script>
+
 </head>
+
 <body>
-<nav class="navbar navbar-default" role="navigation">
-    <!-- Brand and toggle get grouped for better mobile display -->
-    <div class="navbar-header">
-        <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-ex1-collapse">
-            <span class="sr-only">Toggle navigation</span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-        </button>
-        <a class="navbar-brand" href="index.php">HngI7</a>
-    </div>
+  <section class="jumbo">
+  <?php include('fragments/site_header.php'); ?>
+  </section>
 
-    </div><!-- /.navbar-collapse -->
-</nav>
+  <div class="form-area">
+
+    <form class="form-container" method="post">
+
+      <h2 class="login-heading">Login as an Admin</h2>
+      <?php
+      if (!empty($errors)) {
+
+        echo $errors;
+      }
+      if(isset($_GET['getmess'])){
+        echo $_GET['getmess'];
+      }
+
+      if (isset($_GET["blocked"])) {
+        echo  '<p style="margin: 5px; padding: 5px 10px 5px 10px; background: #F7CFCF; color: #6A0E0D; width: 80% !important; text-align: center;">Your account has been de-activated, Please contact a super admin.</p>';
+      }
+
+      ?>
+      <p id="result"></p>
+      <div class="inputWithIcon">
+      <i class="fa fa-envelope" aria-hidden="true"></i>
+      <input type="email" name="email" id="email" placeholder="enter your Email" required />
+      </div>
+
+      <div class="inputWithIcon">
+        <i class="fa fa-key" aria-hidden="true"></i>
+        <input type="password" name="password" id="password" placeholder="enter your password" required>
+      </div>
+      
+      <button type="submit" onclick="signup()" name="login">LOG IN</button>
+      <a href="forgotpassword.php" class="text-danger small-text">Forgot Password?</a>
+    </form>
+  </div>
 
 
-<section id="login">
-    <div class="container">
-        <div class="row">
-            <div class="col-md-6 col-md-offset-3">
-                <div class="panel panel-success">
-                    <div class="panel-heading">Account Login</div>
-                    <div class="panel-body">
-                        <form action="" method="post">
-                            <?php
-                            if(isset($_SESSION['err_msg'])) {
-                                echo $_SESSION['err_msg'];
-                                unset($_SESSION['err_msg']);
-                            }
-                            ?>
-                            <div class="form-group">
-                                <label for="email">Email Address</label>
-                                <input type="email" id="email" name="email" class="form-control" required placeholder="Email Address" value="<?php echo @$_POST['email'];?>">
-                            </div>
-
-                            <div class="form-group">
-                                <label for="password">Password</label>
-                                <input type="password" name="password" required class="form-control" id="password" placeholder="Password">
-                            </div>
-
-
-                            <div class="form-group">
-                                <input type="submit" name="login" class="btn btn-success" value="Login">
-                            </div>
-
-                        </form>
-                    </div>
-                </div>
-            </div>
+  <footer>
+    <img src="https://res.cloudinary.com/jaycodist/image/upload/v1570722444/hng-brand-logo_gnplmq.svg">
+    <nav>
+      <section>
+        <h2 class="skyblue-text">Quick Links</h2>
+        <div id="link-list">
+          <a href="join-intern.php" class="skyblue-text">Join HNG</a>
+          <a href="index.php" class="skyblue-text">About HNG</a>
+          <a href="donate.html" class="skyblue-text">Become a Sponsor</a>
+          <a href="MentorSetUpPage2.php" class="skyblue-text">Sign up as Mentor</a>
+          <a href="faq.php" class="skyblue-text">FAQ</a>
         </div>
-    </div>
-</section>
+      </section>
+      <section id="contact-section">
+        <h2 class="skyblue-text">Contact Us</h2>
+        <div>
+          <a href="tel:+2348123456789">
+            <strong>
+              Phone: <br />
+            </strong>
+            +234 812 345 6789
+          </a>
+          <br />
+          <a href="mailto:interns@hng.tech">
+            <strong>
+              Email: <br />
+            </strong>
+            interns@hng.tech
+          </a>
+        </div>
+      </section>
+      <section>
+        <h2 class="skyblue-text">Office Address</h2>
+        <p id="address">
+          3 Birrel Avenue <br /> Sabo, Yaba, <br /> Lagos state
+        </p>
+      </section>
+      <section>
+        <h2 class="skyblue-text">Follow Us</h2>
+        <div id="socials">
+          <a href="https://twitter.com/hnginternship" title="Follow on Twitter!"><img src="https://res.cloudinary.com/jaycodist/image/upload/v1570722900/twitter-logo_m1mgzi.svg"></a>
+          <a href="https://www.facebook.com/HNG-Internship-115577009820117/"><img title="Follow on Facebook!" src="https://res.cloudinary.com/jaycodist/image/upload/v1570722900/facebook-logo_bw1hal.svg"></a>
+          <a href="https://dribble.com"><img title="Follow on Dribble!" src="https://res.cloudinary.com/jaycodist/image/upload/v1570722900/dribble-logo_w4vwuz.svg"></a>
+        </div>
+      </section>
+    </nav>
+    <p class="center-text darkblue-text">&copy 2019, HNG Internship. All rights reserved.</p>
+  </footer>
 </body>
+
 </html>
